created controller and service for journey

+++ Project1/Data/Journey.cs
@@ -19,5 +19,5 @@ public partial class Journey
 
     public double? CoveredDistance { get; set; }
 
-    public int? Duration { get; set; }
+    public double? Duration { get; set; }
 }
